requirements.txt
# ============================================================================
# Feishu Agent - 智能知识库助手
# 依赖包列表
# ============================================================================

# ----------------------------------------------------------------------------
# 核心依赖（必须）
# ----------------------------------------------------------------------------

# Web框架和HTTP请求
flask>=2.3.0
requests>=2.31.0

# 配置管理
python-dotenv>=1.0.0

# Redis客户端（用于分布式会话和消息去重）
redis>=4.5.0

# MCP客户端（用于外部工具调用）
mcp>=1.0.0

# ----------------------------------------------------------------------------
# RAG 知识库依赖
# ----------------------------------------------------------------------------

# LangChain 核心和组件
langchain>=0.1.0
langchain-community>=0.0.10      # 文档加载器
langchain-text-splitters>=0.0.1   # 文本分割器
langchain-huggingface>=0.0.2      # HuggingFace嵌入模型
langchain-chroma>=0.0.3          # Chroma向量存储集成

# 向量数据库
chromadb>=0.4.0

# 文本嵌入模型
sentence-transformers>=2.2.0

# HuggingFace模型和工具
huggingface-hub>=0.16.0

# 文档处理
pypdf>=3.17.0                    # PDF解析
unstructured>=0.10.0             # 多格式文档解析

# 文本处理工具
numpy>=1.24.0
tqdm>=4.65.0                     # 进度条

# ----------------------------------------------------------------------------
# AI模型接口
# ----------------------------------------------------------------------------

# DeepSeek API 客户端（通过requests实现，无需额外SDK）
# 注意：本项目直接使用requests调用DeepSeek API
# 如需使用官方SDK，可添加：deepseek>=0.1.0

# ----------------------------------------------------------------------------
# 开发工具（可选）
# ----------------------------------------------------------------------------

# 代码格式化
black>=23.0.0

# 测试框架
pytest>=7.0.0

# 类型提示检查
mypy>=1.0.0

# 代码质量检查
flake8>=6.0.0

# ----------------------------------------------------------------------------
# 系统工具（根据操作系统选择安装）
# ----------------------------------------------------------------------------

# Windows系统（用于文件类型检测）
# python-magic-bin>=0.4.14

# macOS/Linux系统（用于文件类型检测）
# python-magic>=0.4.27
# 或安装系统包：brew install libmagic (macOS) / apt-get install libmagic1 (Ubuntu)

# ----------------------------------------------------------------------------
# 版本说明
# ----------------------------------------------------------------------------
# 测试环境：Python 3.8-3.11
# 主要测试平台：macOS (M1/M2/M3芯片兼容)
# 生产建议：使用虚拟环境或Docker容器